requirements.txt
Flask==3.0.3
gunicorn
python-dotenv>=1.0.0
